move ActiveStore class into its own file and rename corresponding tests
