+++ src/runtime/storage/storage-proxy.ts
@@ -17,7 +17,8 @@ import {ParticleExecutionContext} from '../particle-execution-context.js';
 import {ChannelConstructor} from '../channel-constructor.js';
 import {EntityType, Type} from '../../types/lib-types.js';
 import {Handle, HandleOptions} from './handle.js';
-import {ActiveStore, ProxyMessage, ProxyMessageType, StorageCommunicationEndpoint, StorageCommunicationEndpointProvider} from './store-interface.js';
+import {ProxyMessage, ProxyMessageType, StorageCommunicationEndpoint, StorageCommunicationEndpointProvider} from './store-interface.js';
+import {ActiveStore} from './active-store.js';
 import {Ttl} from '../capabilities.js';
 
 /**
