+++ src/runtime/storage/tests/storage-proxy-test.ts
@@ -11,7 +11,8 @@
 import {assert} from '../../../platform/chai-web.js';
 import {CRDTSingletonTypeRecord, SingletonOperation, SingletonOpTypes} from '../../../crdt/lib-crdt.js';
 import {StorageProxy, NoOpStorageProxy} from '../storage-proxy.js';
-import {ActiveStore, ProxyMessageType} from '../store-interface.js';
+import {ProxyMessageType} from '../store-interface.js';
+import {ActiveStore} from '../active-store.js';
 import {MockHandle, MockStore} from '../testing/test-storage.js';
 import {EntityType, SingletonType} from '../../../types/lib-types.js';
